- chatbot will print LLM generated query - fixes: current page dataset distribution process logic should be moved to `SQLConsoleLoader` (as it's always loaded)
